+++ packages/editor/js/extensions/libs/size/compatibility/height.js
@@ -19,6 +19,7 @@ export function heightFromWPCompatibility({
 
 	switch (blockId) {
 		case 'core/image':
+		case 'core/spacer':
 		case 'core/post-featured-image':
 			if (attributes?.height !== undefined) {
 				attributes.blockeraHeight = attributes?.height;
@@ -63,6 +64,7 @@ export function heightToWPCompatibility({
 			};
 
 		// A string attribute for width with unit
+		case 'core/spacer':
 		case 'core/post-featured-image':
 			if ('reset' === ref?.current?.action) {
 				return {