+++ packages/blocks/core/CHANGELOG.md
@@ -1,8 +1,21 @@
 ## Unreleased
 
+### New Features:
+- Added `Spacer Block` support by Blockera. [[🔗 Link](https://community.blockera.ai/feature-request-1rsjg2ck/post/spacer-block-support-pritFhuc8gbsXko)]
+- Added support for the `Navigation Link Block` by Blockera. [[🔗 Link](https://community.blockera.ai/feature-request-1rsjg2ck/post/supporting-blocks-inside-navigation-block-MIcY979kIVCxkvU)]
+- Added support for the `Home Link Block` by Blockera. [[🔗 Link](https://community.blockera.ai/feature-request-1rsjg2ck/post/supporting-blocks-inside-navigation-block-MIcY979kIVCxkvU)]
+- Added support for the `Submenu Block` and its inner blocks by Blockera. [[🔗 Link](https://community.blockera.ai/feature-request-1rsjg2ck/post/supporting-blocks-inside-navigation-block-MIcY979kIVCxkvU)]
+
+
 ### Bug Fixes
 
 - Fixed an issue with test assertions.
+
+### Automated Tests:
+- Tests for the `Spacer Block` to ensure is supported by Blockera.
+- Added automated test for `Navigation Link Block` to check that the `Back to Parent` navigation buttons work correctly.
+- Added automated test for `Home Link Block` to check that the `Back to Parent` navigation buttons work correctly.
+- Added automated test for `Submenu Block` to check that the `Back to Parent` navigation buttons work correctly.
 
 ### Miscellaneous